services/auth-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.4
python-jose[cryptography]==3.3.0
passlib==1.7.4
email-validator==2.2.0
redis>=5.0.0
fastapi-limiter>=0.1.6
bcrypt==4.1.3
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
passlib==1.7.4
